requirements.txt
mlflow==2.13.0
catboost==1.2.5
imbalanced-learn==0.12.2
scikit-learn==1.4.2
azure-storage-file-datalake==12.15.0
azureml-sdk==1.56.0
